Consolidate array flatten functions
