--- server/src/util/flatten.ts
@@ -1,3 +1,0 @@
-export function flattenArray<T>(nestedArray: T[][]): T[] {
-  return nestedArray.reduce((acc, array) => [...acc, ...array], [])
-}
